btc-notify: make tests start at block height 21 to sidestep bip34 height bug
Then fix the cascading set of test failures. This was a rabbit hole but it
turns out that in general we cannot make *any* assertions about what the first
block subscription event will be when we subscribe. On some level this makes
sense but is still annoying. We need to fix this by actually providing a cursor
when subscribing. That's a future commit though...