btc-notify: take cursor argument in subscribe_blocks
This code will need to be revisited.

If we want to gate events based on the subscription's cursor, then
we need to figure out how it should behave when the cursor is not
unidirectional. Right now, the rest of the system only acts off of
buried blocks, however if we set the cursor at N, we wouldn't want
to swallow an event for BlockStatus::Uncled at N-1 because an
uncled block would result in state changes by the consumer. This
makes me wonder if we should have separate subscription types for
Buried, and "Tip", where the Buried events advance monotonically,
and the Tip events can go either way. Generally speaking, Uncling
can only happen if we have a block to replace it with (the only
exception to this is 'invalidateblock' which is never used outside
of testing). This means that we should still imminently progress
the block height we are processing and sending. For now I am going
to swallow any event that is earlier than the cursor, but it does
mean that it will impair reorg resistance.