btc-notify: alter semantics of subscribe_blocks cursor
This commit changes things so that the cursor semantics are no
longer a "filter" but rather a prefix trim. This will later be
altered again when we implement retro-sync but for now this
cleans things up and paves the way for a lazier approach which
should prevent memory accumulation issues.