feat: improve restart resilience of the bridge node (#218)
* fix(duty-tracker): use max deposit idx of all active contracts as offset

* fix(duty-tracker): persist all contract states post-block-processing atomically

* feat(p2p-service): add ouroboros req channel for self-nagging

* multi: make the ouroboros system mediate all access to the p2p network

This change makes it such that we route all messages and message
requests through the ouroboros (OB) channel system.

Prior to this change messages were sent over the MessageHandler which
sent messages to the OB channel and the p2p network at the "same time".
This means that in practice we would send messages to the p2p network
prior to us fully committing all changes that result from that message
to disk. This creates a nasty edge case where if a node crashes after
it sends the message to the network and before it processes the effects
of its own message, it would result in that nodes state falling behind
its peers.

This reveals what I believe to be a latent requirement: the network
should never become aware of an outbound message before our internal
state is fully synchronized to incorporate the effects of that message.

To accomplish this we remove the `P2PHandle` from the `MessageHandler`
entirely. All messages are sent only to the OB channel. From there in
the event handler for the OB channel, we process that message. Only
if it succeeds, do we forward it to the rest of the network.

As a corollary, we forward all `GetMessageRequest`s through the OB
request channel that we use for "self-nags". In the handler for the
OB message requests, we peel off the ones meant for us and emit the
necessary duties, and otherwise we forward the request to the rest
of the network.

* chore(p2p,duty-tracker): update docs and log-level

---------

Co-authored-by: Keagan McClelland <keagan.mcclelland@gmail.com>