feat: Connector trait
All connector outputs implement this trait. Most of the taproot internals are automatically implemented.

The TaprootWitness struct represents a generic Taproot witness. I realize that this clashes with the TaprootWitness from primitives/scripts/taproot.rs. I feel the latter is ill-named or should be removed, since the "witness" doesn't contain any signature data. I plan to make a follow-up PR that revises the struct and that updates the affected Musig2 signing machinery. Let's discuss this.