refactor: Test connectors with BitcoinNode struct
This struct makes it easy to play transactions on Bitcoin Core. The RPC
calls are abstracted away, to reduce code duplication. The coinbase
outpoints are an easy way to fund transactions. The wallet automatically
signs its inputs.

This commit temporarily disables the code separator and cpfp tests. I
will refactor these tests in the following commits.