feat: PresignedTx trait
This is the direct replacement of the CovenantTx trait in the existing
tx graph crate. The trait makes use of the new SigningInfo. ExtraWitness
is required for 2 transactions: Contest (watchtower signature) and
Watchtower Counterproof (operator signatures). We can discuss the
modelling in the PR. The current code is the most generic, but passing
&() for extra_witness is annoying.