refactor: set sequence of each txin
Automatically set the sequence number of each txin that spends a
connector. This forces the implementor to think about which spend path
is being used and it ensures that the sequence numbers are always up to
date. This commit actually fixes a bug: The sequence of vin[0] of the
slash transaction wasn't set correctly. The system works.